fix(P1-05): 保存事务跨进程加锁;版本校验补全;JSONValue 大整数保真
审查(Request changes,1 P1 + 2 P2)三处修正。

## P1:check-then-rename 窗口不锁,并发写入仍会丢

save 读盘校验到 rename 之间没有锁,两个进程都能通过校验、后写者
覆盖先写者 —— rename 只保证单次替换没有半成品,管不到时序。

修正:FileSystem 增加 withExclusiveLock(flock,锁在 url + ".lock"
稳定旁路上,从不被 rename;进程崩溃内核自动释放),保存事务整体
锁内执行。契约测试新增互斥(8 线程 × 64 次,计数必须不丢)与
「锁内替换目标」两用例;ProjectStore 新增带 barrier 的并发双写:
恰好一个成功,无锁实现 20 轮全部命中竞态(探针实测)。

## P2-1:当前版本可写未来 schema,随即把自己锁成只读

保存只查盘上版本,不查待存 envelope。补 envelope.schemaVersion
必须等于 current,否则 envelopeVersionMismatch;写旧/新版格式
留给显式导出 / 迁移 API。

## P2-2:未知大整数被误报损坏

JSONValue 全部数字走 Double,未来版本的 9007199254740993(2^53+1)
进 unknownFields 后解码丢精度、重编码字节变了,完好的新版快照
被报 corrupted。新增 .integer(Int64) case(Bool → Int64 → Double
的解码顺序),整数字面量原样往返;.integer 与 .number 不相等是有意的。
P1-01 已知限制 2 的精度边界从 2^53 推到 Int64 上限。

## 测试证据

- 228 tests / 27 suites,Debug 与 Release 都过,九条检查全绿;
- 契约测试新增 2 用例(互斥计数、锁内替换),双实现参数化;
- ProjectStore 新增 3 用例(并发双写 20 轮、版本不匹配含 v2/v99、
  大整数未知字段);
- 注入(探针口径 函数/用例):删事务锁 → 1/60 红(20 轮全命中);
  删版本校验 → 1/4 红;JSONValue 丢 Int64 分支 → 7/9 红。
- 测试自身两处竞态被抓:DispatchGroup enter 放线程闭包内、
  主线程 wait 落空 —— enter 移到派生侧。

## 回滚方式

git revert 本提交。envelope 落盘格式对当前版本不变(.integer 只影响
未知字段与 payload 的解码表示,本次是首次产生此类数据的版本)。

+++ Tests/PilotCoreTests/DataFormat/DataEnvelopeTests.swift
@@ -112,7 +112,7 @@ struct DataEnvelopeTests {
         #expect(decoded.unknownFields["futureFlag"] == .bool(true))
         #expect(decoded.unknownFields["futureNote"] == .null)
         #expect(decoded.unknownFields["futureConfig"] == .object([
-            "retries": .number(3),
+            "retries": .integer(3),
             "labels": .array([.string("a"), .string("b")]),
         ]))
